First Commit
This is a Spring Employee Crud where you can save your data and also can delete your data. This project is based on the MVC model where it has Employeedao.java which is used to retrieve the data from the database , Employee.java is used to connect the database and HibernateUtil.java which simplifies the development of Java application to interact with the database.